Split out tests into logical groupings.
These groupings are run in a random order by the deployandtest
script. There is also a new soak test script which runs the
tests ten times in a row in the hope of shaking things out.

It would be cool to be able to run these in parallel, but a
lot of the tests make assumptions about the state of the cloud
before and after than need to be sorted out first. This is
just the first step.