Work in progress JWT authentication.
This patch implements the server side, client library changes,
and the command line client changes. It does not implement the
required changes in the ansible modules or the deployment tooling.

That is, this patch breaks installs but should have all the
functional code there. I've just run out of time to work on this
tonight and I wanted to show the general direction.