Make image locks per node
Images are local to a node not global, therefore locks can be per node.

Set lock timeout to the longest download (and longest HTTP timeout).
The client will wait no longer than if it's request had initiated the
download. So no harm to block waiting for the actual download to finish.