Don't create network during new().
For a start, we might be the network node and therefore shouldn't
enqueue a task. Additionally, creating the network here seems racy
for reasons I can't really explain, so let's just wait until we use
it and see if that helps. I also changed the variable name to
something which doesn't clash with the module name (although it
shouldn't be an issue here).

FIxes #577.