Turn snapshot requests into task
Right now, snapshot aren't tasks. This is a problem since they are
definitely a long lived operation which we can reasonably expect to
time out.

A new field has been added to snapshots, `status`, to track whether or
not they're creating or ready for use. This is perhaps too ad-hoc and
will probably need to change once there's a more robust story around
snapshots.

The API now returns the a UUID for a snapshot immediately instead of
waiting for the operation to finish, since my (possibly wrong)
assumption is that it's very rare for a user to want to immediately
want to use a snapshot that they've made.