Images now have a state.
Use this state to mark images as deleted if they are missing from
disk on startup (fixes #607), but also use them to track if an
image has been downloaded yet. This isn't perfect because it
ignores the transcoded derivatives of an image and doesn't clean
up all users of a missing image, but its a start.

Should we expose image delete in the API? I feel like we should
just be "reference counting" images and clean them up when they
have no users and we are constrained for space, but perhaps
there is a use case for delete?