Fix Floating IPs.
Floating IPs have never worked right. I've had a few goes at getting
the routing and packet mangling right to be able to get traffic
from the outside into the network namespace where the virtual
machines are hiding, but with not much luck.

Then I had a brain wave, and realised each virtual IP can just be a
veth that gets traffic into the namespace for that one IP. There is
some packet mangling, which lives on the inner side of the veth,
but not much. It is important for the floating IP to be on the
inner side of the veth as well, which was confusing to me for a bit.

Fixes #257.