Move interface float / defloat to jobs and stop leaking floating IPs.
While these are quick to execute, I need them as jobs so that I can
queue cleanup tasks from the hypervisor node to the network node
when an instance is deleted.

This means we reap floating IPs from interfaces when the instance
is deleted, and that we also free the floating_gateway when a
network is deleted. For good luck, we also now correctly cleanup
an IPManager when the associated network is deleted.

Fixes #256. Fixes #675.