Remove gluster.
Gluster was never sufficiently reliable, and ultimately this was a
failed experiment which just adds complexity to our codebase. Therefore
remove it. I think there might be a future for a shared block filesystem
in the future, but there is no one currently asking for it and we'll
deal with that when the time comes.

Fixes #835.
Fixes #742.
Fixes #731.
Fixes #728.