Handle vxlan races for vxlan leak detection.
When we start an instance, we create the networks first, and then
the VM. This causes the vxlan leak detection code to briefly trigger,
even though we haven't actually leaked a vxlan. Resolve that by
using a more expensive interface lookup if the in-instance cache
has not yet been populated.

Additionally, only warn of extra vxids when they have been present
for more than five minutes.