Ensure floating gateway IPs are reserved.
I have observed a cluster with a network where the floating IP was
not reserved in the IPManager. This went about as well as you'd
expect when that IP is then reused for a floating IP. I am unsure
if this was a random crash or a bug, but add a safety check for it
until we get to the root cause.