Retrieve artifact metadata with non-existent blob
When a new blob is added to an artifact ie. a
snapshot, the blob UUID is stored in the artifact
before the blob is created in the DB. This causes
an error when retrieving non-existent metadata for
an artifact.

A future fix should resolve this by only storing UUID's
in the DB that refer to objects already in existence.