Do not persist blob reference images.
At the moment we create artifacts of type image when we need to
refer to a specific blob -- for example starting an instance. This
is actually pretty annoying as a user because you end up seeing a
lot of artifacts like this in the artifact list:

   sf://blob/6c3aa525-5af5-4c9e-8371-3549e88e2de8

On the other hand its nice as a developer, because it means I can
treat a blob as just another artifact when starting an instance.

So instead of making someone unhappy, let's just have the concept
of an artifact which isn't persisted to the database. I get my nice
abstraction, and the database stays cleaner.