Chasing nvme_boot CI test unreliability.
Theory: NVMe disks require more IO because we can't do a COW layer
for the instance disks there. Therefore, we need to copy around
hundreds of megabytes, if not gigabytes, of data before the instance
can boot. The CI test for NVMe boot is consistently flakey and when
it fails it shows an empty console log. I think this likely indicates
that the instance disks are still being created and that slowness
is occuring because there's too much IO happening on the test VMs
at that time.

So, let's see what happens if we wait for the console log to have
any data at all (i.e. boot has started) before we start looking for
our expected output.