Ensure cluster wide maintenance is only happening once.
At the moment every node attempts to perform the cluster wide
maintenance tasks as part of its cleaner loop. This is ok with smaller
clusters, but on bigger clusters it means that we're running cleanup
N times more often than needed, where N is the number of nodes in the
cluster.

Implement a simple election scheme using a lock to ensure that only
one daemon is doing cleanup at a time.