Move away from hardcoded trusts for just "system".
The system namespace shouldn't be special, it should be an example
of how namespaces can trust each other. Probably for now we should
default to every namespace trusting system, but you should be able
to trust other namespaces. This change removes the hardcoding, but
does not provide an API to extend trust to another namespace. That
will come next.