Reduce prune frequency.
I didn't realise that doing a prune would incur a complete re-write
of the sqlite database in the form of a journal file. This seems
like the sort of thing we should not be doing every second or so
on a machine where we don't want the SSD to wear out.