Disallow etcd locking while producing an artifact external view.
This is part of chasing why listing relatively small numbers of
artifacts is slow slow it causes API timeouts. It turns out getting
the artifact objects from etcd isn't the slow bit, its producing
the external views for the artifacts. Are we holding lots of locks
or something?

Hopefully progresses #1974.