Be consistent with how we handle agent ready states.
CI defines the agent being ready as any agent_state starting with
"ready". These API calls insisted the agent state be exactly "ready".

The gap is that because some of these tests do a hard reboot,
sometimes the instance comes back up in a "ready (degraded)" state
because systemd is telling us that something was sad after the
unplanned reboot. We should accept that condition as ready, because
the agent works and might be the mechanism to un-degrade the
instance.