Drop the old etcd3 client.
This means we wont be pinned to such an old grpc implementation.
It does imply that we do compactions a little bit differently, but
its not too bad.

At the same time, stop fetching metadata we're not using from etcd.

All of this is because it looks like #2242 might be related to etcd
load...