Calculate extra checksums if we're not in a hurry.
We've been verifying blobs with sha512 for a while, but it would
be nice to have sha1 and xxh128 hashes as well. sha1 is useful for
OVA support, and xxh128 is faster in return for not being
cyptographically valid and therefore might be a nice future option
for blob verification.

Also introduce the concept a queued blob verification, which I
think will be useful for other things later.