Add multiple delivery off events.
Add a gRPC call which allows for the delivery on one event to the
records of multiple objects. This is interesting because sometimes
an event relates to more than one object (a blob transcode for an
instance for example), and you really want to be able to see that
event without guessing what it relates to.