Rewrite etcd locking.
So first off this is part of the ongoing removal of etcd3gw. However,
while doing this work I realised that the leasing aspect of cluster
wide locks wasn't really doing anything for me apart from driving
complexity -- the leases need to be refreshed, but guard against a
lock leaking. However, we basically never leak locks and already
have a loop in the cleaner which reaps any leaks that do occur. So,
if I drop the leases I get simpler code and it shouldn't change
reliability at all.

So that's nice.